refactor: fix signed/unsigned comparison warnings

+++ src/gamelogic/rapidXML/RapidXMLAdapter.cpp
@@ -153,7 +153,7 @@ namespace GameLogic {
     string RapidXMLAdapter::get_entity_texture( size_t id )
     {
         for ( xml_node<> * object_node = this->_object_group_node->first_node( "object" ); object_node; object_node = object_node->next_sibling() ) {
-            if ( std::stoi( object_node->first_attribute( "id" )->value() ) == id) {
+            if ( (size_t)std::stoi( object_node->first_attribute( "id" )->value() ) == id) {
                 xml_node<> * property_texture = object_node->first_node( "properties" )->first_node( "property" );
                 return property_texture->first_attribute( "value" )->value();
             }
